Updated logo, now mirrored on imgur.
Logo has been changed for a number of reasons, scalability being by far the biggest concern. The smaller radius on rounded corners became hard, blurry corners when scaled down just as the speech bubble became a blurry circle and the keyhole/person became hard to see. The logotype is now Din Next, the same typeface used on tox.im and the threads on /gd/ and /g/. I also created a proper .ico file for the program itself, though the 16x16 still needs a bit of work.