Move extra libtool options into configure
This should allow to keep the libtool options all in one place and at
the same time define different options depending on the host.

Made sure that -no-undefined is set only on Win32. Although no side
effects on Linux and OSX have been observed so far, it's probably better
to play it safe; it does not seem to be needed/does not seem to matter on *nix,
only required for Win32.