Fix regression that disabled shared library builds
My previous attempt that was supposed to disable shared library builds
with nacl had a side effect which basically disabled shared libs for all
configurations.

Eventhough AC_DISABLE_SHARED was used inside an if clause it seemed to
take over in any case.

I could not find a clean way around this, so had to override internal
libtool variables. Will check with the libtool people regarding a
cleaner implementation.