Fixed typo.
This typo doesn't seem to cause any issue because according to the
code in vanilla NaCl the first crypto_box_ZEROBYTES (for encryption)
and crypto_box_BOXZEROBYTES (for decryption) of the array passed to
the crypto_box*() functions don't need to be zero for it to work.

The documentation however clearly states that they need to be zero
which means they need to be zero.