Make Tox more portable
(u)intN_t is defined **optional** by the C standard and you are using it in places that it is not needed.

(u)int_fastN_t and (u)int_leastN_t is the portable way. 

For decent targets you could definitely get away with this. It isn't the perfect alternative, but it's a pretty good start.

Really though I'd probably wrap everything in a test macro to handle weird systems that don't care about full ISO compliance (never minding POSIX). But that's up to you.

The main idea here is portability.

This is needed for some systems, such as the HP-UX.

HP-UX currently can't use Tox.