Expose constants as functions.
These functions simply return the constants. They are a stable ABI, so that if
constants change, the ABI of these functions won't. Code solely relying on these
functions will remain compatible with future values of those constants.

The functions are currently not exposed in tox.h, because this is pending a
change in apidsl to generate accessors for "const" values.