feat: Let CMake create the dll instead of doing so ourselves
We used to make the dll manually as we previously had 3 dlls:
libtoxcore.dll, libtoxav.dll and libtoxencryptsave.dll, but for Windows
we wanted them to be all combined into libtox.dll with all the
dependencies included: libsodium, libopus, libvpx, pthreads, etc, to
reduce the overall dll size and simplify linking. However, since CMake
now produces a single libtoxcore.dll with toxcore, toxav and
toxencryptsave included, we don't have to do this manually anymore.

This results in the dll being named libtoxcore.dll instead of the
libtox.dll that it previously was, matching the static libtoxcore.a's
name.