refactor: Observers/ignored peers can now send and receive custom packets
The Observer role was intended to prevent peers from being
disruptive and/or interacting with other peers in the group.
It wasn't intended to cripple custom protocols running on-top
of the groups such as file sharing and message syncing.

In cases where it might be undesirable for observers to use
custom packets (e.g. starting a file transfer) the client will
have the ability to decide whether or not to allow it.